+++ web/src/web.ts
@@ -1,2 +1,5 @@
-export * from './baseHelp.component';
-export * from './help.service';
+export * from './misc/baseHelp.component';
+export * from './services/help.service';
+export * from './misc/utils';
+export * from './modules/markdown.module';
+export * from './directives/renderMarkdown/renderMarkdown.directive';